expert_system.py added: First line contains the shebang (/usr/bin/env search for a python3 interpreter to read and interpret the file)
Then comes a `if __name__ == '__file__:'` block. It is a good practice to put your entrypoint in such block.
If you then execute this file and use it as an entrypoint, __name__ variable will have '__main__' for value.

In many other cases with python development, if you want to put a little unit test to test a module (== python file)
you can put it in such if blocks. Otherwise if you put it in global scope your test will be executed each time you
import your module from an other file.

pass keyword is just here to avoid letting the if block empty (so there is no syntax error).